feat: Building out tabs and FE calls
Lots of things happening here: initially I added the fe calls for the rover data on the back end, and though it is not reflected, I went through several versions of thinking about how to structure the calls and properly update the state. I then built out the initial tabs implementation. So far all of this is theoretical, but I am now ready to test.