feat: Full functionality, bonus features
Now the app is fully functional in the way expected from the spec. I can see all the relevant data for each rover on click, and the photo is also clickable to choose a different photo, and theres an extra call to choose a different random day for photos. I moved the initial rover call into the page load, since it should function like a componentDidMount/useEffect call and update the state when the page loads. I still see plenty of room for improvement, possibly styling or immutability, id also like the active tab to be shown.